+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/view/ViewTest.kt
@@ -35,6 +35,7 @@ class ViewTest {
         assertTrue(view.hasFocus())
     }
 
+    @Suppress("Deprecation")
     @Test
     fun `visibility helper methods`() {
         val view = TextView(testContext)